Pending changes exported from your codespace

+++ src/test/java/com/jpmc/midascore/UserPopulator.java
@@ -1,9 +1,10 @@
 package com.jpmc.midascore;
 
-import com.jpmc.midascore.component.DatabaseConduit;
-import com.jpmc.midascore.entity.UserRecord;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import com.jpmc.midascore.component.DatabaseConduit;
+import com.jpmc.midascore.entity.UserRecord;
 
 @Component
 public class UserPopulator {
@@ -17,7 +18,7 @@ public class UserPopulator {
         String[] userLines = fileLoader.loadStrings("/test_data/lkjhgfdsa.hjkl");
         for (String userLine : userLines) {
             String[] userData = userLine.split(", ");
-            UserRecord user = new UserRecord(userData[0], Float.parseFloat(userData[1]));
+            UserRecord user = new UserRecord(userData[0], Double.parseDouble(userData[1]));
             databaseConduit.save(user);
         }
     }
